Update README.md
Money market funds are part of a class of assets known as collective investment schemes. These schemes assemble funds from various investors and act as general agents who manage and generate returns that are in turn redistributed proportionally to the investors. Money market funds invest in money market instruments that are characterized as risk free and short termed.